+++ frontend/src/lib/components/PropertyFilters/components/PropertyFiltersDisplay.tsx
